perf: vm interpreter safety fixes + performance optimizations
Safety:
- Replace stack-allocated bc_buf[64KB] with sys_mmap() dynamic allocation
- Unified goto cleanup + sys_munmap() on all exit paths (prevent leak)
- Pure syscall wrappers (no libc dependency)

Performance:
- Computed goto dispatch table (replace switch-case, ~20-30% speedup)
- XOR decryption 8-byte widening (u64 bulk xor, ~8x decrypt speed)
- addr_map insertion sort + binary search for BR_REG (O(log n))

Fixes:
- Field-level copy instead of struct assignment (avoid implicit memcpy under -nostdlib)
- Loop fill dispatch table instead of [0...255] range init (same memcpy issue)

Tested: both original_arm64.vmp and rsa_verify_demo.vmp pass on device

+++ stub/vm_handlers/h_mem.h
@@ -3,6 +3,7 @@
  *
  * LDRB / LDR(32/64) / STRB / STR(32/64)
  * 编码: [op | d/base | n/src | offset16]  共 5B
+ * offset16 按 signed int16 解释以支持 LDUR/STUR 负偏移
  */
 #ifndef H_MEM_H
 #define H_MEM_H
@@ -10,13 +11,12 @@
 #include "../vm_decode.h"
 #include "../vm_types.h"
 
-
 /* ---- 加载 (LDR) ---- */
 
 /* LDRB Wd, [Xn, #off16] */
 static inline u32 h_load8(vm_ctx_t *vm) {
   u8 d = vm->bc[vm->pc + 1], n = vm->bc[vm->pc + 2];
-  u16 off = rd16(&vm->bc[vm->pc + 3]);
+  i16 off = (i16)rd16(&vm->bc[vm->pc + 3]);
   vm->R[d & 31] = *(u8 *)(vm->R[n & 31] + off);
   return 5;
 }
@@ -24,7 +24,7 @@ static inline u32 h_load8(vm_ctx_t *vm) {
 /* LDR Wd, [Xn, #off16] */
 static inline u32 h_load32(vm_ctx_t *vm) {
   u8 d = vm->bc[vm->pc + 1], n = vm->bc[vm->pc + 2];
-  u16 off = rd16(&vm->bc[vm->pc + 3]);
+  i16 off = (i16)rd16(&vm->bc[vm->pc + 3]);
   vm->R[d & 31] = *(u32 *)(vm->R[n & 31] + off);
   return 5;
 }
@@ -32,7 +32,7 @@ static inline u32 h_load32(vm_ctx_t *vm) {
 /* LDR Xd, [Xn, #off16] */
 static inline u32 h_load64(vm_ctx_t *vm) {
   u8 d = vm->bc[vm->pc + 1], n = vm->bc[vm->pc + 2];
-  u16 off = rd16(&vm->bc[vm->pc + 3]);
+  i16 off = (i16)rd16(&vm->bc[vm->pc + 3]);
   vm->R[d & 31] = *(u64 *)(vm->R[n & 31] + off);
   return 5;
 }
@@ -42,7 +42,7 @@ static inline u32 h_load64(vm_ctx_t *vm) {
 /* STRB Wn, [Xb, #off16] */
 static inline u32 h_store8(vm_ctx_t *vm) {
   u8 b = vm->bc[vm->pc + 1], n = vm->bc[vm->pc + 2];
-  u16 off = rd16(&vm->bc[vm->pc + 3]);
+  i16 off = (i16)rd16(&vm->bc[vm->pc + 3]);
   *(u8 *)(vm->R[b & 31] + off) = (u8)vm->R[n & 31];
   return 5;
 }
@@ -50,7 +50,7 @@ static inline u32 h_store8(vm_ctx_t *vm) {
 /* STR Wn, [Xb, #off16] */
 static inline u32 h_store32(vm_ctx_t *vm) {
   u8 b = vm->bc[vm->pc + 1], n = vm->bc[vm->pc + 2];
-  u16 off = rd16(&vm->bc[vm->pc + 3]);
+  i16 off = (i16)rd16(&vm->bc[vm->pc + 3]);
   *(u32 *)(vm->R[b & 31] + off) = (u32)vm->R[n & 31];
   return 5;
 }
@@ -58,7 +58,7 @@ static inline u32 h_store32(vm_ctx_t *vm) {
 /* STR Xn, [Xb, #off16] */
 static inline u32 h_store64(vm_ctx_t *vm) {
   u8 b = vm->bc[vm->pc + 1], n = vm->bc[vm->pc + 2];
-  u16 off = rd16(&vm->bc[vm->pc + 3]);
+  i16 off = (i16)rd16(&vm->bc[vm->pc + 3]);
   *(u64 *)(vm->R[b & 31] + off) = vm->R[n & 31];
   return 5;
 }
